pyproject.toml
[tool.poetry]
name = "fastapiblog"
version = "0.1.0"
description = ""
authors = ["andrejmatveev <andrey240101@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.1"
uvicorn = "^0.16.0"
SQLAlchemy = "^1.4.29"
databases = "^0.5.3"
psycopg2-binary = "^2.9.2"
asyncpg = "^0.25.0"
python-multipart = "^0.0.5"
python-jose = "^3.3.0"
passlib = "^1.7.4"
email-validator = "^1.1.3"
alembic = "^1.7.5"
bcrypt = "^3.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
